YEXT_project/sourcing-source-pre-processing/requirements.txt
autopep8==1.6.0
pycodestyle==2.8.0
python-dotenv==0.20.0
toml==0.10.2
apscheduler==3.9.1
paramiko==2.11.0
black==22.6.0
ijson==3.1.4
requests-toolbelt==0.9.1
paramiko==2.11.0
pysftp==0.2.9
xmltodict==0.12.0